STM32U5_AWS_QuickConnect/requirements.txt
pyyaml
botocore
boto3==1.22.0
requests
pyserial
cryptography
black
pre-commit
pyasn1
cbor2
click
jinja2
imgtool==1.9.0